LayoutTests/editing/pasteboard/contenteditable-pre-2-live-range-expected.txt

This test copies and pastes content inside pre that is an editing host. CyberKit should not clone pre.
To manually test, cut and paste "hello\nworld" CyberKit should not nest pre (no red borders).

Before cut paste:
| <pre>
|   "<#selection-anchor>hello"
|   <br>
|   "world<#selection-focus>"

After cut paste:
| <pre>
|   "hello"
|   <br>
|   "world<#selection-caret>"
